Commonize some interfaces
