+++ common/data.ts
@@ -1,0 +1,12 @@
+export
+interface IMessage {
+  text: string;
+  user: {
+    name: string;
+  };
+}
+
+export
+interface IUser {
+  name: string
+}
